Criado as validações de entradas

+++ Consultorio.Domain/Entity/Paciente.cs
@@ -2,7 +2,7 @@
 {
     public class Paciente : Usuario
     {
-        public double? Altura { get; set; }
-        public double? Peso { get; set; }
+        public decimal? Altura { get; set; }
+        public decimal? Peso { get; set; }
     }
 }
